backend/signal-generator/requirements.txt
fastparquet>=2024.11.0,<2026
flask>=3.0,<4
numpy>=2.0,<3
pandas>=2.2,<4
scikit-learn>=1.5,<2
lightgbm>=4.5,<5
google-cloud-firestore>=2.0,<3
google-cloud-storage>=2.0,<3
google-cloud-pubsub>=2.0,<3
mlflow-skinny>=2.0,<3
